requirements.txt
aiohttp==3.5.4
aiosocks==0.2.6
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Automat==0.6.0
blinker==1.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
configparser==4.0.2
constantly==15.1.0
cryptography==2.1.4
discord==1.0.1
discord.py==1.2.5
distro-info===0.18ubuntu0.18.04.1
google-images-download==2.8.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
idna-ssl==1.1.0
imgurpython==1.1.7
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
multidict==4.6.1
netifaces==0.10.4
numpy==1.17.2
oauthlib==2.0.6
PAM==0.4.2
Pillow==6.2.1
praw==6.4.0
prawcore==1.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.22.0
requests-unixsocket==0.1.5
rule34==1.6.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.13.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing-extensions==3.7.4.1
ufw==0.36
unattended-upgrades==0.1
update-checker==0.16
urllib3==1.25.8
Wand==0.5.7
websocket-client==0.56.0
websockets==6.0
yarl==1.3.0
youtube-dl==2020.1.24
zope.interface==4.7.1
